# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pet Parents API Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://3.132.243.209/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/settings": {
            "get": {
                "tags": [
                    "Settings"
                ],
                "summary": "settings",
                "operationId": "settings",
                "responses": {
                    "200": {
                        "description": "settings",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SettingResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/schemas/BadRequest"
                    },
                    "404": {
                        "$ref": "#/components/schemas/Notfound"
                    },
                    "500": {
                        "$ref": "#/components/schemas/Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "BadRequest": {
                "title": "Bad Request",
                "description": "BadRequest",
                "properties": {
                    "dsfsfds": {
                        "title": "dsfsfds",
                        "description": "Bad Request",
                        "type": "string"
                    }
                },
                "type": "string"
            },
...
            "Forbidden": {
                "title": "Forbidden",
                "description": "Forbidden",
                "properties": {
                    "forbidden": {
                        "title": "forbidden",
                        "description": "Forbidden",
                        "type": "string"
                    }
                },
                "type": "string"
            },
            "Notfound": {
                "title": "Not found",
                "description": "Not found",
                "properties": {
                    "notfound": {
                        "title": "Not found",
                        "description": "Not found",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "SettingResponse": {
                "title": "settings Response",
                "description": "settings Response",
                "properties": {
                    "data": {
                        "type": "object",
                        "example": {
                            "site_title": "dfdfdf",
                            "site_email": "test@gmail.com",
                            "phone": "1212121",
                            "logo": null,
                            "facebook": null,
                            "youtube": null,
                            "insta": null,
                            "linkedin": null,
                            "copyright": null
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            },
            "Token": {
                "type": "apiKey",
                "name": "Token",
                "in": "header",
                "scheme": "Token"
            }
        }
    },
...
    "security": {
        "0": [],
        "passport": {
            "type": "oauth2",
            "description": "Laravel passport oauth2 security.",
            "in": "header",
            "scheme": "https",
            "flows": {
                "password": {
                    "authorizationUrl": "http://127.0.0.1:8000//oauth/authorize",
                    "tokenUrl": "http://127.0.0.1:8000//oauth/token",
                    "refreshUrl": "http://127.0.0.1:8000//token/refresh",
                    "scopes": []
                }
            }
        }
    }
}